refactor: extract SwapForm logic into useSwapForm (#185)
SwapForm.vue held ~499 script LOC (asset list derivation, the
wallet-driven protocolFilter init, route quoting with its debounce, swap
validation and the multi-wallet submit path) in one SFC, against the
house rule that a template be a thin projection of already-computed
values.

Move that logic verbatim into a co-located useSwapForm() composable;
SwapForm.vue becomes template plus composable wiring. Behaviour is
unchanged and the existing SwapForm.test.ts (13 cases) passes untouched.

The four module-scope vars (time, route, timeOut, id) move inside the
composable, matching the CloseDialog split. This is behaviour-neutral for
a single dialog instance and removes shared mutable state across
instances: route and the debounce timer are written and read only within
one instance lifecycle, and id only indexes wallet.history[id] guards
that never match a Date.now() key.

No subcomponent is split out — the template has no separable preview
block. Continues the #185 SFC split.
